enable groupby over arrow only on spark 2.3 or newer
